ide_rule_library/requirements.txt
# IDE Rule Library Dependencies
# Install with: pip install -r requirements.txt

# Google Gemini API client
google-generativeai>=0.3.2

# Neo4j driver
neo4j>=5.16.0

# Environment variables
python-dotenv>=1.0.1

# Retry logic with exponential backoff
tenacity>=8.2.3

# YAML configuration parsing
pyyaml>=6.0

# Configuration validation
pydantic>=2.0.0
